requirements.txt
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
idna==2.10
jsonpath-rw==1.4.0
ply==3.11
psycopg2==2.8.6
python-decouple==3.3
requests==2.24.0
six==1.15.0
urllib3==1.25.10
mysql-connector-python==8.0.25